backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
jose
pydantic
whisper
Pillow
authlib
starlette
httpx
bleach
jinja2
